Script create/delete instance profiles.
When an EC2 instance is created, in order for CodeDeploy
to work with it it must have certain permissions setup
and have the CodeDeployAgent installed. These scripts try
to automate this.

create-env.sh now requires an instance profile name to be specified.
There is now no default instance profile in the scripts.

Removed superfluous service principals from policy.
We don't need to have any permissions specified for non eu-west-1 zones.