Don't use the docker group
Docker has a number of privilege escalation issues [1]. This means that
adding a user to the docker group allows them to run arbitrary commands
as root, without the auditing and control that a tool such as `sudo`
enforces.

Given this, we should require use of `sudo` to access docker, rather
than using the `docker` group to control access.

[1]: http://www.projectatomic.io/blog/2015/08/why-we-dont-let-non-root-users-run-docker-in-centos-fedora-or-rhel/