remove instance profile from create-application code
In order to have a CodeDeploy application, you need a Service Role ARN
to add to your deployment group.  However, that Role doesn't need to be
attached to an Instance Profile.  Moreover, we can hardcode the value of
the Service Role ARN because it's the same each time.

This also adds a script to create the code deploy service role.