don't renew a certificate unless we have a local copy
You can't run `lego renew` unless you have a local copy of the cert.  I
can't work out what the condition `when: not cert_file.stat.exists` was
for, because it can't have worked.

This also updates the computed command string to use the `>-` folded
string indicator in order to strip the trailing newline from the string.