Specify instance root volume size for indexer
In 63502ce86, we switched to the minimal AMI image. This had the
unintended side effect of reducing the root volume size everywhere to
2Gb, because when you don't specify the root volume size it seems to
default to the size of the AMI (or at least this is my theory).

As a result, we ended up unable to deploy the indexer because the root
filesystem was full.

Explicitly setting the root volume size fixes this.

Note that due to hashicorp/terraform#3941, this won't automatically
cause the instance to be reprovisioned.

Also note that this commit only changes the indexer instance, but all
instances are affected.