modify ansible script to create databases for new merged app
FIXME: we need to work out how to do this better.  The old code
connected to RDS from the indexer server, because there was exactly one
such server per environment, and it was able to connect to both RDS
instances.  There is no such convenient server in the infrastructure for
the new app.