Use native lists where possible
Based on the migration guide: https://www.terraform.io/upgrade-guides/0-7.html#migrating-to-native-lists-and-maps

We had a few modules which output list-type values as join()ed strings,
as this was all you could do in terraform 0.6.  Version 0.7 introduces
native lists as a type.  This converts some old usages to the new native
list type.

Things *not* done:

 - the mint or presentation modules, as their days are numbered
 - the variables which are configured through .tfvars, which will need
   more care