Default all subnets to /24 in new environments
When we generate the .tfvars file for a new environment, we create quite
a few /27 subnets (address space size: 32).  This has caused problems in
the past, in particular because ELBs use a surprisingly large number of
addresses and don't want to be put into a subnet with fewer than 10 free
addresses.

I've upped all subnets to be /24-sized (address space size of 256).
It's hard to understand the ipsubnet() function so I did a test run on
the 192.0.0.0/16 block.  Beforehand we allocated these subnets:

   - public cidr blocks:
     - 192.0.10.0/27
     - 192.0.10.32/27
     - 192.0.10.64/27
   - openregister cidr blocks:
     - 192.0.21.0/24
     - 192.0.22.0/24
     - 192.0.23.0/24
   - openregister db cidr blocks:
     - 192.0.20.0/27
     - 192.0.20.32/27
     - 192.0.20.64/27

After this change, we allocate these subnets instead:

   - public cidr blocks:
     - 192.0.1.0/24
     - 192.0.2.0/24
     - 192.0.3.0/24
   - openregister cidr blocks:
     - 192.0.21.0/24
     - 192.0.22.0/24
     - 192.0.23.0/24
   - openregister db cidr blocks:
     - 192.0.24.0/24
     - 192.0.25.0/24
     - 192.0.26.0/24

I tried to avoid reusing the same IPs in old-form vs new-form subnets.