Use AWS Managed NAT Gateway
We still need to tidy up the NAT instance so that it doesn't have two
roles (Bastion and NAT Gateway). It seems sensible to do these as two
separate deploys.

We are also only going to add a NAT instance per-AZ later as this
requires having to add more route tables and associations which seems
sensible to do separately too.