Redirect job-centre to jobcentre
This adds a new module, `redirect`, which allows you to redirect all
requests from one domain to another.  It's built from an S3 static
website (which does the actual redirecting), a CloudFront
distribution (which is just there to support HTTPS) and a Route 53
record.

It also instantiates the module for the job-centre discovery register,
redirecting to the jobcentre register instead.