move register_diocese to new module/register module
It feels icky to repeatedly instantiate the same three modules for each
separate register.  So I've created a new module `register` which does
the job all in one go.  This commit also modifies register_diocese.tf to
use the new module, but none of the others.

I've already applied this in discovery, but not in alpha (which also has
a diocese register).