move route 53 record out of module/register temporarily
While we're migrating registers from the old modules to the new module,
we should handle the route 53 record separately.  This will allow us to
create new registers, deploy the app to them, ensure they're working,
and finally flip the route 53 record from the old to the new, before
destroying the old infrastructure.