Enable CloudTrail
This enables CloudTrail logging from all regions to a single bucket.

The Terraform code to enable CloudTrail comes from a repo that David
King has been working on. We've been playing with CloudWatch Alarms that
are triggered from CloudTrail logs to see if we can usefully alert on
potential security issues. At the moment we're not sure if these will be
useful (as we haven't tested them with CloudTrail logs from a "real"
account), which is why all the alarm names are currently prefixed with
"test".