Use temporary aliases to avoid collisions in CloudFront
Terraform complains when trying to create multiple CloudFront
distributions that respond to the same alias (for example, it's not
possible to create two distributions that both respond to the FQDN
`register.register.gov.uk`). In order to move to the new CloudFront
distributions we'll need to coordinate DNS changes and the removal of
the old CloudFront distributions. So that amount of time we're running
without a CDN is as small as possible it seems sensible to spin up the
distributions (which takes a long time) and then edit them to be correct
(by deploying a revert of this commit) later.