Undo work to automate CloudFront distributions
It turns out that doing this migration without any downtime is more
involved/fiddly that we'd hoped.

We had originally planned to temporarily run the beta registers without
any CDN so that we could migrate from the manually configured
distributions to the new Terraform managed distributions. Running
without the CDN is only required because CloudFront does not allow you
to run multiple, separate distributions that respond to the same host.
However, it is not possible to run `*.register.gov.uk` registers without
CloudFront because this is where TLS termination happens.

Given that the multi-tenancy work is happening soon we've decided to
shelve this until after that because this will allow us to have one
CloudFront distribution and it'll be simpler to migrate.

This reverts commit 7329b3fe7020bee864bde8498a68ad1fe0221453 and 15d2fa54adbd93357a4d200b680a9f2f3ad2772d.