Upgrade to Amazon Linux 2016.09.1
Amazon Linux has rolling releases:

> The Amazon Linux AMI repository structure is configured to deliver a
> continuous flow of updates that allow you to roll from one version of
> the Amazon Linux AMI to the next.
>
> ...
>
> In other words, Amazon Linux AMIs are treated as snapshots in time, with
> a repository and update structure that provides the latest packages that
> we have built and pushed into the repository.
Source: https://aws.amazon.com/amazon-linux-ami/faqs/

This means that it should just mean we have less packages to upgrade
when we spin up a new instance. However, I'll apply it just to to the
multi-tenanted groups only as we'll be trashing the single-tenanted
groups soon.