Create DNS record resources individually
This works around the problems we were having with creating DNS records
using a single Terraform `aws_route53_record` resource which iterated
over a list of records in a list.

We had problems with doing this, due to the way Terraform matches local
state addresses (eg `module.multi.aws_route53_record.multi.4`) against
remote provider IDs (in this case a R53 record which is uniquely
identified by a tuple of the Hosted Zone ID, record name and record
type). When reordering items in a list that is consumed by a Terraform
resource that builds R53 records all records that have a local Terraform
address that no longer matches what it expects to see remotely will
first be deleted and then created again. These two operations are not
atomic.

Given the above Terraform behaviour this means we have to create each
individual R53 DNS record as a separate Terraform resource. Terraform
provides no way of doing this in a DRY way which means we end up with a
big collection of related, virtually identical DNS records that have
been declared individually.

There is some additional ugliness here which means that each register
must now be part of the same group in every environment. Only the
address register will be affected by this. This was done because
Terraform conditionals evaluate both sides of a branch and therefore it
doesn't appear possible to consume the output of a module (in our case
this is where the information about the load balancer the record points
at comes from) that possibly doesn't have populated outputs. For
example, this causes an error in Terraform if `module.a` doesn't return
any output:

```
"${false ? module.a.record_fqdn : module.b.record_fqdn}"
```