Set the TTL for NS delegation of subdomains on openregister.org
I think the right thing was trying to be achieved here (ie trying to
keep the NS delegation for `{test, discovery, ...}.openregister.org`
short so it can be changed easily). However, the TTL here is for the
NS record in the wrong zone.

NS records at the point-of-delegation are non-authoritative. It is the
TTL of the NS record in the authoritative zone that is respected:

>  Defined in RFC 1035. NS RRs appear in two places. Within the zone file,
>  in which case they are authoritative records for the zone's name
>  servers. At the point of delegation for either a subdomain of the zone
>  or in the zone's parent in which case they are non-authoritative. Thus,
>  the zone example.com's parent zone (.com) will contain
>  non-authoritative NS RRs for the zone example.com at its point of
>  delegation (point of delegation is the term frequently used to describe
>  the NS RRs in the parent that delegate a zone of subdomain) and
>  subdomain.example.com will have non-authoritative NS RRS in the zone
>  example.com at its point of delegation. NS RRs at a point of delegation
>  are never authoritative only NS RRs within the zone are regarded as
>  authoritative. While this may look a fairly trivial point, is has
>  important implications for DNSSEC.

Source: http://www.zytrax.com/books/dns/ch8/ns.html