Explicitly specify the RDS maintenance window
Since Terraform 0.8.3 it appears the behaviour of the `aws_db_instance`
resource has changed when passing an empty string to the
`maintenance_window` argument. We only pass an empty string because the
module that contains `aws_db_instance` exposes this as a configurable
option.

The `maintenance_window` argument on `aws_db_instance` is an optional
argument, which is chosen by Amazon if it is not specified. Since we
never specified a maintenance window I've decided to just set the value
in the `*.tfvar` files for each environment to the value Amazon selected
for us when the RDS instance was created.