Set SOA negative cache TTL to one hour
Without this a `NXDOMAIN` response will be cached by recursors for 24
hours. Given we're hoping to add registers fairly regularly (which will
require new subdomains) it seems sensible to keep this short in order to
avoid the case where someone accidentally(?) tries to resolve a new
register before the records have been created.

The `name-server` field is being set to the first name server in the
list of name servers the hosted zone uses. This is likely to be
different from the value created by AWS, however, it should be
acceptable to set it to any name server that responds authoritatively
for the domain

See: http://www.zytrax.com/books/dns/ch8/soa.html