migrate some of our instances to ubuntu
This migrates some of our instances to ubuntu.  Specifically, the
even-numbered instances, when numbering from 0 (so where we have 3
instances, the first and last instance will be migrated to ubuntu).

This only migrates some of the instances so that we don't cause
downtime through this migration.  Once this has been deployed
everywhere, we can go back and migrate the rest.

Our user data (in users.yaml) needed very little changes to support
this.  All I did was:

 - rename some packages which are named differently on ubuntu (aws-cli
   to awscli, docker to docker.io)
 - remove yum-cron and its config (we have a card to reinstate
   ubuntu's equivalent)

I've preserved the original user data as
modules/register_group/old-users.yaml, to support keeping the old
Amazon Linux instances around.

Once this is merged, the remaining tasks are:

 - deploy this everywhere
 - migrate the bastion host
 - migrate the remaining register_group instances
 - clean up the interim code (ie remove old-users.yaml and the
   `element(list(...), count.index)` hacks in this commit)