manage our own cloudtrail
it's not possible to have two instances of the aws-security-alarms
module, because they conflict with each other.

It's not clear that it's valuable for us to do this anyway; we didn't
need most of the functionality and it's probably simpler to manage the
S3 bucket and cloudtrail ourselves.

This works with an appropriate `terraform state mv` command to shift
the S3 bucket and cloudtrail resources.