Remove Ansible playbook to create Terraform `*.tfvars`
This seems like overkill for our use case. We so rarely create
environments that it doesn't feel like having to maintain this script
with any additional variables/configuration we add is worth it.

It might mean that we need to spend a bit more time keeping our
`*.tfvars` consistent across environments, however this seems preferable
to having to use Ansible in order to this.

This also gets us slightly closer to getting rid of Ansible entirely.