Remove unused Terraform
This looks like a half-completed attempt at using Terraform to manage
the deployment/configuration of CodeDeploy. At some point we should move
to using Terraform to manage CodeDepoy (we can then remove the shell
scripts at the root of this repo), but at the moment the stuff here
appears to be unused.

Even if this were used we should probably just use the AWS the following
managed policy:

`arn:aws:iam::aws:policy/service-role/AWSCodeDeployRole`