Stop performing reverse DNS lookups of RDS clients
It turns out this was causing slowness for our RDS clients. This was
most pronounced when deploying but almost certainly affected most DB
connections in some way.

In a normal situation you would expect these reverse lookups to be fast.
However, because we accidentally chose "private" IP ranges within
`172.0.0.0/8` but outside of `172.16.0.0/12` we ended up assigning
privately IPs in the public address space. This means when Postgres
performs a reverse DNS lookup we rely on the owners of those IP address
to have configurations that respond with timely responses.

Frequently this is unproblematic. I imagine in some instances (although
I've not observed this personally) we'll see correct, affirmative
responses. In other case we've seen `NXDOMAIN` responses which should be
cached based off the `SOA` record. The slow ones I've personally
observed have responded with `SERVFAIL`; these are slow and are probably
not cached by recursors (§7.1 of RFC2308) so the failure has a bigger
impact.

Ultimately, we should look to fix the issue of misallocated private IP
ranges, but in the short-term this workaround should stop us from seeing
problems.