Demote registers from beta to alpha
- registration-district
- notifiable-animal-disease
- notifiable-animal-disease-confirmation-category
- notifiable-animal-disease-investigation-category

We do not usually do this but these registers were prematurely promoted to beta.
registry-office did not have approval from the custodian to be promoted to beta.
The notifiable-animal-disease registers contained some errors that existed in alpha
and were not fixed before promotion to beta. There will be a postmortem written
for this incident so that we can make improvements to the alpha -> beta promotion process.