Reduce the number of redundant/useless metrics sent to InfluxDB
Dropwizard provides a lot of metrics by default. Some of these are not
very valuable because they can be easily derived at query time (the mean
rates). Some of these seem excessive for our use case (it seems good
enough just to track p99 and p95).

We are only doing this so that we can keep the cost of InfluxCloud down.