Add Lambda to send availability to Performance Platform
This calculates a mean uptime for the `/records` endpoint for all
registers currently in Beta. It then forwards this value to the
Performance Platform.

NB: Although the documentation on this is unclear the `Uptime` percent
value returned by StatusCake is the percent uptime over a 24 hour
period. This was confirmed by talking to their support team.