Temporarily limit Terraform's parallelism
Statuscake is fronted by Cloudflare and because there's some
rate-limiting there it means that we can easily trigger a rate-limiter
from that results in lot of 429 responses.

This change makes Terraform `plan` and `apply` slow, but reliable. I
tried larger numbers than `1` but these also triggered the rate-limiter.

We've been in contact with Statuscake and they're looking to try and up
the rate limit. However, it possibly also should be the case that the
Terraform provider should limit appropriately. This change should only
exist for as long as one of these things still haven't been resolved.