Remove outdated and stale scripts
These were written a long time ago and were probably ran once(?) in
order to bootstrap an AWS account. Given that happens so rarely (if
ever) I think it doesn't make sense to keep these around.

We really should reimplement what they do in Terraform but for the time
being I think having them linger around in the root of this repository
is confusing.