Remove `address` in test
It's a large register which we currently don't handle well. Migrations
are slow on this register and can cause deployments to appear to fail.
We can reassess whether it is sensible to have a register of this size
in test if/when we plan to create a register of this size.