Send uptime metrics as starting from the previous day
The Performance Platform expects the metric for a period to have a
timestamp for the start of that period. In this case the period is a day
and the uptime from StatusCake should be for the previous 24 hours.

We send our metrics at UTC midnight so the start of this period is
actually the previous day.