Enable deletion of basic registers in alpha and discovery
This means it will be possible to delete the contents of the `register`,
`field` and `datatype` registers. These two environments/phases are not
expected to be consistent or satisfy the append-only rules of further
environments/phases. It is acceptable for these registers to be emptied
and reloaded as part of iterating on the shape of registers.