Fix canonicalization to remove blank values from metadata
Items associated with user entries are currently run through canonicalization, but items associated with system entries are not. This means that when we create an RSF file to load data for a particular register, the metadata entries will not match those for the environment we're trying to load to, as the hashes will be different between the canonicalized item (taken from the environment) and the non-canonicalized item we're loading.

This commit also updates expected test data to remove all empty register properties.