Forward `X-Forwarded-Host` header to origin
We currently set and forward this header on the CloudFront distributions
that front `*.register.gov.uk`. We require it in order to return HTML
that is styled appropriately for pages on `gov.uk`.

Whilst migrating to the PaaS we will be in a situation where we have two
CloudFront distributions in front of our PaaS apps for registers on
`register.gov.uk`. It will look something like this for
`country.register.gov.uk`.

```
country.register.gov.uk (CloudFront) -> *.beta.openregister.org (CloudFront) -> cloudapps.digital (PaaS)
```

We therefore need to ensure the `X-Fowarded-Host` header set by the
`country.register.gov.uk` CloudFront distribution is passed on by the
CloudFront distribution that fronts `*.beta.openregister.org` to PaaS.