Reduce CloudFront TTL to 60 seconds for *.openregister.org
As part of migrating to PaaS we introduced new CloudFront
distributions for the *.openregister.org domains. This means that
when we make updates to discovery and alpha environments the cache
does not invalidate for 24 hours. This is temporary solution so that
we can see updates immediately. In the near future we should do some
work to make our cache headers more intelligent. The register.gov.uk
domain will continue to cache for 24 hours.