Add average response time as metric to performance platform data
This commit adds two new metrics - `responsetime` (from the `avgresponse` provided by Pingdom) and `location` - which will be sent alongside the existing payload to the Performance Platform. Splitting out `location` from the name of the check will make filtering easier, whilst `responsetime` is added to as additional data to display on the report.