Remove terrafrom code and config for influxcloud
We used to use InfluxCloud to store dropwizard metrics from
openregister-java when it was hosted on AWS EC2. When we moved
to PaaS we never reimplemented this feed in favour of using a
Reliability Engineering provided solution. Although we don't have
that yet, any work we do to reimplement metrics from ORJ should be
in conjunction with Reliability Engineering's tool that we use for
our Ruby application. We therefore no longer need this code.