Add another script to update titles and descriptions at the same time
This is equivalent to running update_descriptions.py and then
name-a-register.sh.

At the moment, this updates one register at a time. This is a bit
tedious and for updating multiple registers so if I run it again I might
change it to take a CSV and generate multiple updates at once.